+++ sshmirror/__init__.py
@@ -11,4 +11,4 @@ __all__ = [
 	'VersionAlreadyExists',
 ]
 
-__version__ = '0.1.18'
+__version__ = '0.1.19'